0.Prep/requirements.txt
torch
protobuf
sentencepiece
pillow
transformers==4.49.0
requests
timm
Flask
sqlalchemy-iris